backend/pyproject.toml
[tool.poetry]
name = "store-backend"
version = "0.1.0"
description = "Backend API for Store - Inventory and Sales Management System"
authors = ["Store Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-ratelimit = "^4.1.0"
django-axes = "^6.1.1"
argon2-cffi = "^23.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-django = "^4.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


